+++ src/newick_parser.h
@@ -1,6 +1,7 @@
 #ifndef NEWICK_PARSER_H
 	#define NEWICK_PARSER_H
 
+  #include <Rcpp.h>
 	#include <string>
 	#include "unrooted_tree.h"
 
@@ -11,8 +12,10 @@
 		public:
 			~NewickParser() {};
 			UnrootedTree *parseStr(string inputStr);
+			UnrootedTree *parseStr(Rcpp::CharacterVector string);
 			UnrootedTree *parseFile(const char* filename);
 			std::vector<UnrootedTree *> parseMultiFile(const char *filename);
+			std::vector<UnrootedTree *> parseMultiStr(Rcpp::CharacterVector string);
 			bool isError();
 
 		private: